renaming the events package. Removing the explicit module. Moving the http client to generic.
