+++ events/events_get_test.go
@@ -1,4 +1,4 @@
-package gomulocity_event
+package events
 
 import (
 	"testing"
