+++ events/events_update_test.go
@@ -1,4 +1,4 @@
-package gomulocity_event
+package events
 
 import (
 	"encoding/json"
