--- events/go.mod
@@ -1,5 +1,0 @@
-module github.com/tarent/gomulocity/event
-
-go 1.14
-
-require github.com/tarent/gomulocity v0.0.0-20200618145517-4672ed43fce1
